requirements.txt
fastapi
uvicorn
sqlalchemy
pillow
requests
jinja2
pytest
httpx